\name{BsMD-package}
\alias{BsMD-package}
\alias{BsMD}
\docType{package}
\title{
Bayes screening and model discrimination follow-up designs
}
\description{
Bayes screening and model discrimination follow-up designs
}
\details{
\tabular{ll}{
Package: \tab BsMD\cr
Type: \tab Package\cr
Version: \tab 2013.0718\cr
Date: \tab 2013-07-18\cr
License: \tab GPL version 2 or later\cr
}
The packages allows you to perform the calculations and analyses described
in Mayer, Stainberg and Box paper in Technometrics, 1996.
}
\author{
Author: Ernesto Barrios based on Daniel Meyer's code.
Maintainer: Ernesto Barrios <ebarrios@itam.mx>
}
\references{
    Box and Mayer, 1986; Box and Mayer, 1993; Mayer, Steinberg and Box, 1996.
}
\keyword{ package }
\examples{
    data(BM86.data)
}
